src/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
clickhouse-connect==0.8.18
pydantic-settings==2.11.0
email-validator==2.3.0
